requirements.txt
flask
pandas
kaggle
sqlalchemy
python-dotenv
